requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.12.3
google-generativeai==0.8.5
python-dotenv==1.2.1
requests==2.32.5
